+++ include/monolish_blas_double.hpp
@@ -27,8 +27,10 @@ void mul(const vector<double> &a, const double alpha, vector<double> &y);
 void div(const vector<double> &a, const vector<double> &b, vector<double> &y);
 void div(const vector<double> &a, const double alpha, vector<double> &y);
 
-void vecadd(const vector<double> &a, const vector<double> &b, vector<double> &y);
-void vecsub(const vector<double> &a, const vector<double> &b, vector<double> &y);
+void vecadd(const vector<double> &a, const vector<double> &b,
+            vector<double> &y);
+void vecsub(const vector<double> &a, const vector<double> &b,
+            vector<double> &y);
 //////////////////////////////////////////////////////
 //  Vector
 //////////////////////////////////////////////////////